+++ draft-packages/tooltip/docs/Tooltip.stories.tsx
@@ -1,11 +1,11 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
-import * as React from "react"
+import React from "react"
 import informationIcon from "@kaizen/component-library/icons/information-white.icon.svg"
 import meatballsIcon from "@kaizen/component-library/icons/meatballs.icon.svg"
 import { Tag } from "@kaizen/draft-tag"
 import { Icon, Paragraph, Heading } from "@kaizen/component-library"
 import { withDesign } from "storybook-addon-designs"
-import { Button, IconButton } from "@kaizen/draft-button"
+import { Button, IconButton } from "@kaizen/button"
 import { Tooltip } from "@kaizen/draft-tooltip"
 import isChromatic from "chromatic/isChromatic"
 import { figmaEmbed } from "../../../storybook/helpers"
@@ -48,7 +48,6 @@ export const DefaultKaizenSiteDemo = props => (
     </Tooltip>
   </div>
 )
-
 DefaultKaizenSiteDemo.storyName = "Default (Kaizen Site Demo)"
 DefaultKaizenSiteDemo.parameters = {
   info: {
@@ -262,6 +261,7 @@ export const StickerSheet = props => (
     </div>
   </div>
 )
+StickerSheet.parameters = { chromatic: { disable: false } }
 
 export const OverflowScroll = props => (
   <>