+++ packages/tabs/docs/Tabs.stories.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react"
 import { Box, Paragraph } from "@kaizen/component-library"
 import { Card } from "@kaizen/draft-card"
-import { Button } from "@kaizen/draft-button"
+import { Button } from "@kaizen/button"
 import { CATEGORIES } from "../../../storybook/constants"
 import { Tabs, TabList, TabPanels, TabPanel, Tab } from "../index"
 
@@ -54,6 +54,7 @@ export const Uncontrolled = () => (
     </TabPanels>
   </Tabs>
 )
+Uncontrolled.parameters = { chromatic: { disable: false } }
 
 export const Controlled = () => {
   const [selectedIndex, setSelectedIndex] = useState<number>(0)
@@ -165,7 +166,6 @@ export const UsageInCard = () => (
     </Tabs>
   </Card>
 )
-
 UsageInCard.parameters = {
   backgrounds: { default: "Gray 100" },
 }