+++ draft-packages/badge/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-badge@1.10.2...@kaizen/draft-badge@1.10.3) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/draft-badge
+
+
+
+
+
+## [1.10.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-badge@1.10.1...@kaizen/draft-badge@1.10.2) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/draft-badge
+
+
+
+
+
+## [1.10.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-badge@1.10.0...@kaizen/draft-badge@1.10.1) (2022-02-28)
+
+**Note:** Version bump only for package @kaizen/draft-badge
+
+
+
+
+
 # [1.10.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-badge@1.9.0...@kaizen/draft-badge@1.10.0) (2022-02-10)