+++ draft-packages/card/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-card@1.7.2...@kaizen/draft-card@1.7.3) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/draft-card
+
+
+
+
+
+## [1.7.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-card@1.7.1...@kaizen/draft-card@1.7.2) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/draft-card
+
+
+
+
+
+## [1.7.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-card@1.7.0...@kaizen/draft-card@1.7.1) (2022-02-28)
+
+**Note:** Version bump only for package @kaizen/draft-card
+
+
+
+
+
 # [1.7.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-card@1.6.0...@kaizen/draft-card@1.7.0) (2022-02-10)