+++ draft-packages/modal/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.4.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@8.4.11...@kaizen/draft-modal@8.4.12) (2022-03-02)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [8.4.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@8.4.10...@kaizen/draft-modal@8.4.11) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [8.4.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@8.4.9...@kaizen/draft-modal@8.4.10) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [8.4.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@8.4.8...@kaizen/draft-modal@8.4.9) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [8.4.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@8.4.7...@kaizen/draft-modal@8.4.8) (2022-02-28)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [8.4.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@8.4.6...@kaizen/draft-modal@8.4.7) (2022-02-28)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
 ## [8.4.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@8.4.5...@kaizen/draft-modal@8.4.6) (2022-02-28)
 
 **Note:** Version bump only for package @kaizen/draft-modal