+++ draft-packages/popover/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@3.4.4...@kaizen/draft-popover@3.4.5) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/draft-popover
+
+
+
+
+
+## [3.4.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@3.4.3...@kaizen/draft-popover@3.4.4) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/draft-popover
+
+
+
+
+
+## [3.4.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@3.4.2...@kaizen/draft-popover@3.4.3) (2022-02-28)
+
+**Note:** Version bump only for package @kaizen/draft-popover
+
+
+
+
+
 ## [3.4.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-popover@3.4.1...@kaizen/draft-popover@3.4.2) (2022-02-28)
 
 **Note:** Version bump only for package @kaizen/draft-popover