+++ draft-packages/tile/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.6.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@5.6.8...@kaizen/draft-tile@5.6.9) (2022-03-02)
+
+**Note:** Version bump only for package @kaizen/draft-tile
+
+
+
+
+
+## [5.6.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@5.6.7...@kaizen/draft-tile@5.6.8) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/draft-tile
+
+
+
+
+
+## [5.6.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@5.6.6...@kaizen/draft-tile@5.6.7) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/draft-tile
+
+
+
+
+
+## [5.6.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@5.6.5...@kaizen/draft-tile@5.6.6) (2022-02-28)
+
+**Note:** Version bump only for package @kaizen/draft-tile
+
+
+
+
+
+## [5.6.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@5.6.4...@kaizen/draft-tile@5.6.5) (2022-02-28)
+
+**Note:** Version bump only for package @kaizen/draft-tile
+
+
+
+
+
 ## [5.6.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@5.6.3...@kaizen/draft-tile@5.6.4) (2022-02-28)
 
 **Note:** Version bump only for package @kaizen/draft-tile