+++ packages/a11y/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.3.5...@kaizen/a11y@1.3.6) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/a11y
+
+
+
+
+
+## [1.3.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.3.4...@kaizen/a11y@1.3.5) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/a11y
+
+
+
+
+
+## [1.3.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.3.3...@kaizen/a11y@1.3.4) (2022-02-28)
+
+**Note:** Version bump only for package @kaizen/a11y
+
+
+
+
+
 ## [1.3.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.3.2...@kaizen/a11y@1.3.3) (2022-02-28)
 
 **Note:** Version bump only for package @kaizen/a11y