+++ packages/brand-moment/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.15](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.9.14...@kaizen/brand-moment@1.9.15) (2022-03-02)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.9.14](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.9.13...@kaizen/brand-moment@1.9.14) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.9.13](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.9.12...@kaizen/brand-moment@1.9.13) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.9.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.9.11...@kaizen/brand-moment@1.9.12) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.9.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.9.10...@kaizen/brand-moment@1.9.11) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.9.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.9.9...@kaizen/brand-moment@1.9.10) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.9.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.9.8...@kaizen/brand-moment@1.9.9) (2022-02-28)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.9.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.9.7...@kaizen/brand-moment@1.9.8) (2022-02-28)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
 ## [1.9.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.9.6...@kaizen/brand-moment@1.9.7) (2022-02-28)
 
 **Note:** Version bump only for package @kaizen/brand-moment