+++ packages/loading-spinner/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/loading-spinner@2.0.1...@kaizen/loading-spinner@2.0.2) (2022-03-01)
+
+**Note:** Version bump only for package @kaizen/loading-spinner
+
+
+
+
+
+## [2.0.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/loading-spinner@2.0.0...@kaizen/loading-spinner@2.0.1) (2022-02-28)
+
+**Note:** Version bump only for package @kaizen/loading-spinner
+
+
+
+
+
 # 2.0.0 (2022-02-18)